third_party/blink/web_tests/platform/generic/external/wpt/svg/geometry/parsing/r-invalid-expected.txt

This is a testharness.js-based test.
FAIL e.style['r'] = "10" should not set the property value assert_equals: expected "" but got "10"
PASS e.style['r'] = "auto" should not set the property value
PASS e.style['r'] = "10px 20px" should not set the property value
PASS e.style['r'] = "-1px" should not set the property value
PASS e.style['r'] = "-10%" should not set the property value
Harness: the test ran to completion.
